pyproject.toml
[project]
name = "xmpd"
version = "1.4.4"
description = "Multi-source music daemon: syncs YouTube Music and Tidal libraries to MPD"
authors = [
    {name = "Tunç Yıldırım", email = "tuncenator@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "ytmusicapi>=1.0.0",
    "pyyaml>=6.0",
    "python-mpd2>=3.1.0",
    "yt-dlp>=2023.0.0",
    "aiohttp>=3.9.0",
    # Unofficial Tidal client. Pin to a known-good range; bump if Tidal's
    # API changes break us. PKCE flow not used (HiRes deferred per
    # PROJECT_PLAN.md Cross-Cutting Concerns > Tidal HiRes Streaming).
    "tidalapi>=0.8.11,<0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pytest-cov>=4.0.0",
    "ruamel.yaml>=0.18,<0.19",
]

[project.scripts]
xmpd = "xmpd.__main__:main"

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["xmpd*"]
exclude = ["tests*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
norecursedirs = ["tests/research"]
markers = [
    "tidal_integration: live tests against the user's real Tidal account; requires XMPD_TIDAL_TEST=1",
]
